Refactor tests into coherent modules
